+++ Source/CudaUtilities.h
@@ -35,9 +35,9 @@ private:
 bool InitializeCuda(void);
 void HandleCudaError(const cudaError_t CudaError, const char* pDescription = "");
 void HandleCudaKernelError(const cudaError_t CudaError, const char* pName = "");
-int GetTotalCudaMemory(void);
-int GetAvailableCudaMemory(void);
-int GetUsedCudaMemory(void);
+size_t GetTotalCudaMemory(void);
+size_t GetAvailableCudaMemory(void);
+size_t GetUsedCudaMemory(void);
 int GetMaxGigaFlopsDeviceID(void);
 bool SetCudaDevice(const int& CudaDeviceID);
 void ResetDevice(void);
